// chess-game/src/app/_/enums/coins.enum.ts
export enum Coins {
  //comman
  Zero = 0,
  Route = 0,

  //black
  _bKing = 1,
  _bQueen = 2,

  _bBishopR = 3,
  _bKnightR = 4,
  _bRookR = 5,

  _bBishopL = 6,
  _bknightL = 7,
  _bRookL = 8,
  // _bBishopL = -3,
  // _bknightL = -4,
  // _bRookL = -5,

  _bPawn1 = 11,
  _bPawn2 = 12,
  _bPawn3 = 13,
  _bPawn4 = 14,
  _bPawn5 = 15,
  _bPawn6 = 16,
  _bPawn7 = 17,
  _bPawn8 = 18,

  //white
  _wKing = 31,
  _wQueen = 32,

  _wBishopR = 33,
  _wKnightR = 34,
  _wRookR = 35,

  // _wBishopL = -33,
  // _wknightL = -34,
  // _wRookL = -35,
  _wBishopL = 36,
  _wknightL = 37,
  _wRookL = 38,

  _wPawn1 = 41,
  _wPawn2 = 42,
  _wPawn3 = 43,
  _wPawn4 = 44,
  _wPawn5 = 45,
  _wPawn6 = 46,
  _wPawn7 = 47,
  _wPawn8 = 48,
}